requirements.txt
Flask>=2.0.0   # For the web framework itself
PyJWT>=2.0.0   # For generating and verifying JSON Web Tokens
bcrypt>=3.2.0  # For hashing passwords
python-dotenv>=0.15.0 # For loading environment variables from a .env file (used in refactored config.py)

# Werkzeug is a dependency of Flask and will be installed automatically.
# sqlite3, datetime, os, uuid, base64, functools are part of the Python standard library.